refactor: 保持明细表和交叉表排序逻辑统一
